refactor(friends): separate feature into components, hooks, types and pages
