+++ web/src/features/friends/types/index.ts
@@ -1,0 +1,10 @@
+export interface Friend {
+	id: number;
+	name: string;
+	avatar: string;
+	level: string;
+	streak: number;
+	isOnline: boolean;
+	isBestFriend: boolean;
+	interests: string[];
+}
